requirements.txt
selenium==4.15.2
pytest==7.4.3
allure-pytest==2.13.2
webdriver-manager==4.0.1

